+++ src/utils.ts
@@ -1,6 +1,6 @@
 export function toString(val: unknown): string {
     let value = String(val);
-    if (value === '[object Object]') {
+    if (value === "[object Object]") {
         try {
             value = JSON.stringify(val);
         } catch {}