+++ test/err.test.ts
@@ -1,8 +1,8 @@
-import { assert } from 'conditional-type-checks';
-import { Err, Ok } from '../src';
-import { eq, expect_never } from './util';
+import { assert } from "conditional-type-checks";
+import { Err, Ok } from "../src";
+import { eq, expect_never } from "./util";
 
-test('Constructable & Callable', () => {
+test("Constructable & Callable", () => {
     const a = new Err(3);
     expect(a).toBeInstanceOf(Err);
     eq<typeof a, Err<number>>(true);
@@ -12,10 +12,10 @@ test('Constructable & Callable', () => {
     eq<typeof b, Err<number>>(true);
 
     function mapper<T>(fn: (val: string) => T): T {
-        return fn('hi');
+        return fn("hi");
     }
     const mapped = mapper(Err);
-    expect(mapped).toMatchResult(new Err('hi'));
+    expect(mapped).toMatchResult(new Err("hi"));
 
     // TODO: This should work!
     // eq<typeof mapped, Err<string>>(true);
@@ -24,7 +24,7 @@ test('Constructable & Callable', () => {
     mapper<Err<number>>(Err);
 });
 
-test('ok, err, and val', () => {
+test("ok, err, and val", () => {
     const err = new Err(32);
     expect(err.err).toBe(true);
     assert<typeof err.err>(true);
@@ -36,13 +36,13 @@ test('ok, err, and val', () => {
     eq<typeof err.val, number>(true);
 });
 
-test('static EMPTY', () => {
+test("static EMPTY", () => {
     expect(Err.EMPTY).toBeInstanceOf(Err);
     expect(Err.EMPTY.val).toBe(undefined);
     eq<typeof Err.EMPTY, Err<void>>(true);
 });
 
-test('else, unwrapOr', () => {
+test("else, unwrapOr", () => {
     const e1 = Err(3).else(false);
     expect(e1).toBe(false);
     eq<false, typeof e1>(true);
@@ -52,51 +52,53 @@ test('else, unwrapOr', () => {
     eq<false, typeof e2>(true);
 });
 
-test('expect', () => {
+test("expect", () => {
     expect(() => {
-        const err = Err(true).expect('should fail!');
+        const err = Err(true).expect("should fail!");
         expect_never(err, true);
-    }).toThrowError('should fail!');
+    }).toThrowError("should fail!");
 });
 
-test('unwrap', () => {
+test("unwrap", () => {
     expect(() => {
-        const err = Err({ message: 'bad error' }).unwrap();
+        const err = Err({ message: "bad error" }).unwrap();
         expect_never(err, true);
     }).toThrowError('{"message":"bad error"}');
 });
 
-test('map', () => {
+test("map", () => {
     const err = Err(3).map((x: any) => Symbol());
     expect(err).toMatchResult(Err(3));
     eq<typeof err, Err<number>>(true);
 });
 
-test('andThen', () => {
-    const err = new Err('Err').andThen(() => new Ok(3));
-    expect(err).toMatchResult(Err('Err'));
+test("andThen", () => {
+    const err = new Err("Err").andThen(() => new Ok(3));
+    expect(err).toMatchResult(Err("Err"));
     eq<typeof err, Err<string>>(true);
 });
 
-test('mapErr', () => {
-    const err = Err('32').mapErr((x) => +x);
+test("mapErr", () => {
+    const err = Err("32").mapErr((x) => +x);
     expect(err).toMatchResult(Err(32));
     eq<typeof err, Err<number>>(true);
 });
 
-test('iterable', () => {
+test("iterable", () => {
     for (const item of Err([123])) {
         expect_never(item, true);
-        throw new Error('Unreachable, Err@@iterator should emit no value and return');
+        throw new Error(
+            "Unreachable, Err@@iterator should emit no value and return",
+        );
     }
 });
 
-test('to string', () => {
-    expect(`${Err(1)}`).toEqual('Err(1)');
-    expect(`${Err({ name: 'George' })}`).toEqual('Err({"name":"George"})');
+test("to string", () => {
+    expect(`${Err(1)}`).toEqual("Err(1)");
+    expect(`${Err({ name: "George" })}`).toEqual('Err({"name":"George"})');
 });
 
-test('stack trace', () => {
+test("stack trace", () => {
     function first(): Err<number> {
         return second();
     }
@@ -112,6 +114,6 @@ test('stack trace', () => {
     expect(err.stack).toMatch(/Err\(1\)/);
     expect(err.stack).not.toMatch(/ErrImpl/);
 
-    const err2 = Err(new Error('inner error'));
+    const err2 = Err(new Error("inner error"));
     expect(err2.stack).toMatch(/Err\(Error: inner error\)/);
 });