+++ test/ok.test.ts
@@ -1,8 +1,8 @@
-import { assert } from 'conditional-type-checks';
-import { Err, Ok, Result } from '../src';
-import { eq, expect_never, expect_string } from './util';
+import { assert } from "conditional-type-checks";
+import { Err, Ok, Result } from "../src";
+import { eq, expect_never, expect_string } from "./util";
 
-test('Constructable & Callable', () => {
+test("Constructable & Callable", () => {
     const a = new Ok(3);
     expect(a).toBeInstanceOf(Ok);
     eq<typeof a, Ok<number>>(true);
@@ -12,11 +12,11 @@ test('Constructable & Callable', () => {
     eq<typeof b, Ok<number>>(true);
 
     function mapper<T>(fn: (val: string) => T): T {
-        return fn('hi');
+        return fn("hi");
     }
 
     const mapped = mapper(Ok);
-    expect(mapped).toMatchResult(new Ok('hi'));
+    expect(mapped).toMatchResult(new Ok("hi"));
 
     // TODO: This should work!
     // eq<typeof mapped, Ok<string>>(true);
@@ -25,7 +25,7 @@ test('Constructable & Callable', () => {
     mapper<Ok<number>>(Ok);
 });
 
-test('ok, err, and val', () => {
+test("ok, err, and val", () => {
     const err = new Ok(32);
     expect(err.err).toBe(false);
     assert<typeof err.err>(false);
@@ -37,13 +37,13 @@ test('ok, err, and val', () => {
     eq<typeof err.val, number>(true);
 });
 
-test('static EMPTY', () => {
+test("static EMPTY", () => {
     expect(Ok.EMPTY).toBeInstanceOf(Ok);
     expect(Ok.EMPTY.val).toBe(undefined);
     eq<typeof Ok.EMPTY, Ok<void>>(true);
 });
 
-test('else, unwrapOr', () => {
+test("else, unwrapOr", () => {
     const e1 = Ok(3).else(false);
     expect(e1).toBe(3);
     eq<number, typeof e1>(true);
@@ -53,44 +53,44 @@ test('else, unwrapOr', () => {
     eq<number, typeof e2>(true);
 });
 
-test('expect', () => {
-    const val = Ok(true).expect('should not fail!');
+test("expect", () => {
+    const val = Ok(true).expect("should not fail!");
     expect(val).toBe(true);
     eq<boolean, typeof val>(true);
 });
 
-test('unwrap', () => {
+test("unwrap", () => {
     const val = Ok(true).unwrap();
     expect(val).toBe(true);
     eq<boolean, typeof val>(true);
 });
 
-test('map', () => {
+test("map", () => {
     const mapped = Ok(3).map((x) => x.toString(10));
-    expect(mapped).toMatchResult(Ok('3'));
+    expect(mapped).toMatchResult(Ok("3"));
     eq<typeof mapped, Ok<string>>(true);
 });
 
-test('andThen', () => {
-    const ok = new Ok('Ok').andThen(() => new Ok(3));
+test("andThen", () => {
+    const ok = new Ok("Ok").andThen(() => new Ok(3));
     expect(ok).toMatchResult(Ok(3));
     eq<typeof ok, Ok<number>>(true);
 
-    const err = new Ok('Ok').andThen(() => new Err(false));
+    const err = new Ok("Ok").andThen(() => new Err(false));
     expect(err).toMatchResult(Err(false));
     eq<typeof err, Result<string, boolean>>(true);
 });
 
-test('mapErr', () => {
-    const ok = Ok('32').mapErr((x: any) => +x);
-    expect(ok).toMatchResult(Ok('32'));
+test("mapErr", () => {
+    const ok = Ok("32").mapErr((x: any) => +x);
+    expect(ok).toMatchResult(Ok("32"));
     eq<typeof ok, Ok<string>>(true);
 });
 
-test('iterable', () => {
+test("iterable", () => {
     let i = 0;
-    for (const char of Ok('hello')) {
-        expect('hello'[i]).toBe(char);
+    for (const char of Ok("hello")) {
+        expect("hello"[i]).toBe(char);
         expect_string(char, true);
         i++;
     }
@@ -105,11 +105,13 @@ test('iterable', () => {
     for (const item of Ok(1)) {
         expect_never(item, true);
 
-        throw new Error('Unreachable, Err@@iterator should emit no value and return');
+        throw new Error(
+            "Unreachable, Err@@iterator should emit no value and return",
+        );
     }
 });
 
-test('to string', () => {
-    expect(`${Ok(1)}`).toEqual('Ok(1)');
-    expect(`${Ok({ name: 'George' })}`).toEqual('Ok({"name":"George"})');
+test("to string", () => {
+    expect(`${Ok(1)}`).toEqual("Ok(1)");
+    expect(`${Ok({ name: "George" })}`).toEqual('Ok({"name":"George"})');
 });