+++ test/rxjs.test.ts
@@ -1,5 +1,5 @@
-import { Err, Ok, Result } from '../src';
-import { Observable, of } from 'rxjs';
+import { Err, Ok, Result } from "../src";
+import { Observable, of } from "rxjs";
 import {
     elseMap,
     elseMapTo,
@@ -13,13 +13,13 @@ import {
     resultSwitchMap,
     tapResultErr,
     tapResultOk,
-} from '../src/rxjs-operators';
-import { eq } from './util';
+} from "../src/rxjs-operators";
+import { eq } from "./util";
 
-const goodVal: Observable<Result<string, number>> = of(Ok('good'));
+const goodVal: Observable<Result<string, number>> = of(Ok("good"));
 const badVal: Observable<Result<string, number>> = of(Err(0));
 
-test('resultMap', () => {
+test("resultMap", () => {
     const goodValMapped = goodVal.pipe(
         resultMap((msg) => {
             eq<typeof msg, string>(true);
@@ -40,7 +40,7 @@ test('resultMap', () => {
     expect(badValMapped).toMatchObsResult(Err(0));
 });
 
-test('resultMapErr', () => {
+test("resultMapErr", () => {
     const goodValMappedErr = goodVal.pipe(
         resultMapErr((num) => {
             eq<typeof num, number>(true);
@@ -48,7 +48,7 @@ test('resultMapErr', () => {
         }),
     );
     eq<typeof goodValMappedErr, Observable<Result<string, Date>>>(true);
-    expect(goodValMappedErr).toMatchObsResult(Ok('good'));
+    expect(goodValMappedErr).toMatchObsResult(Ok("good"));
 
     const badValMappedErr = badVal.pipe(
         resultMapErr((num) => {
@@ -60,7 +60,7 @@ test('resultMapErr', () => {
     expect(badValMappedErr).toMatchObsResult(Err(new Date(0)));
 });
 
-test('resultMapTo', () => {
+test("resultMapTo", () => {
     const goodValMappedTo = goodVal.pipe(resultMapTo(500));
     eq<typeof goodValMappedTo, Observable<Result<number, number>>>(true);
     expect(goodValMappedTo).toMatchObsResult(Ok(500));
@@ -70,17 +70,17 @@ test('resultMapTo', () => {
     expect(badValMappedTo).toMatchObsResult(Err(0));
 });
 
-test('resultMapErrTo', () => {
-    const goodValMappedErrTo = goodVal.pipe(resultMapErrTo('err'));
+test("resultMapErrTo", () => {
+    const goodValMappedErrTo = goodVal.pipe(resultMapErrTo("err"));
     eq<typeof goodValMappedErrTo, Observable<Result<string, string>>>(true);
-    expect(goodValMappedErrTo).toMatchObsResult(Ok('good'));
+    expect(goodValMappedErrTo).toMatchObsResult(Ok("good"));
 
-    const badValMappedErrTo = badVal.pipe(resultMapErrTo('err'));
+    const badValMappedErrTo = badVal.pipe(resultMapErrTo("err"));
     eq<typeof badValMappedErrTo, Observable<Result<string, string>>>(true);
-    expect(badValMappedErrTo).toMatchObsResult(Err('err'));
+    expect(badValMappedErrTo).toMatchObsResult(Err("err"));
 });
 
-test('elseMap', () => {
+test("elseMap", () => {
     const goodValElseMapped = goodVal.pipe(
         elseMap((num) => {
             eq<typeof num, number>(true);
@@ -89,7 +89,7 @@ test('elseMap', () => {
     );
     eq<typeof goodValElseMapped, Observable<string | Date>>(true);
 
-    expect(goodValElseMapped).toMatchObs('good');
+    expect(goodValElseMapped).toMatchObs("good");
 
     const badValElseMapped = badVal.pipe(
         elseMap((num) => {
@@ -102,11 +102,11 @@ test('elseMap', () => {
     expect(badValElseMapped).toMatchObs(new Date(0));
 });
 
-test('elseMapTo', () => {
+test("elseMapTo", () => {
     const goodValElseMapped = goodVal.pipe(elseMapTo(new Date(100)));
     eq<typeof goodValElseMapped, Observable<string | Date>>(true);
 
-    expect(goodValElseMapped).toMatchObs('good');
+    expect(goodValElseMapped).toMatchObs("good");
 
     const badValElseMapped = badVal.pipe(elseMapTo(new Date(100)));
     eq<typeof badValElseMapped, Observable<string | Date>>(true);
@@ -114,14 +114,17 @@ test('elseMapTo', () => {
     expect(badValElseMapped).toMatchObs(new Date(100));
 });
 
-test('switchMap', async () => {
+test("switchMap", async () => {
     const goodValSwitchMappedToResult = goodVal.pipe(
         resultSwitchMap((a) => {
             eq<typeof a, string>(true);
             return of(Ok(new Date(3)) as Result<Date, string>);
         }),
     );
-    eq<typeof goodValSwitchMappedToResult, Observable<Result<Date, number | string>>>(true);
+    eq<
+        typeof goodValSwitchMappedToResult,
+        Observable<Result<Date, number | string>>
+    >(true);
     expect(goodValSwitchMappedToResult).toMatchObsResult(Ok(new Date(3)));
 
     const goodValSwitchMappedToBadResult = goodVal.pipe(
@@ -130,7 +133,10 @@ test('switchMap', async () => {
             return of(Err(new Date(3)) as Result<number, Date>);
         }),
     );
-    eq<typeof goodValSwitchMappedToBadResult, Observable<Result<number, Date | number>>>(true);
+    eq<
+        typeof goodValSwitchMappedToBadResult,
+        Observable<Result<number, Date | number>>
+    >(true);
     expect(goodValSwitchMappedToBadResult).toMatchObsResult(Err(new Date(3)));
 
     const badValSwitchMappedToResult = badVal.pipe(
@@ -139,7 +145,10 @@ test('switchMap', async () => {
             return of(Ok(new Date(3)) as Result<Date, string>);
         }),
     );
-    eq<typeof badValSwitchMappedToResult, Observable<Result<Date, number | string>>>(true);
+    eq<
+        typeof badValSwitchMappedToResult,
+        Observable<Result<Date, number | string>>
+    >(true);
     expect(badValSwitchMappedToResult).toMatchObsResult(Err(0));
 
     const goodValSwitchMappedToNonResult = goodVal.pipe(
@@ -148,7 +157,9 @@ test('switchMap', async () => {
             return of(new Date(3));
         }),
     );
-    eq<typeof goodValSwitchMappedToNonResult, Observable<Result<Date, number>>>(true);
+    eq<typeof goodValSwitchMappedToNonResult, Observable<Result<Date, number>>>(
+        true,
+    );
     expect(goodValSwitchMappedToNonResult).toMatchObsResult(Ok(new Date(3)));
 
     const badValSwitchMappedToNonResult = badVal.pipe(
@@ -157,18 +168,23 @@ test('switchMap', async () => {
             return of(new Date(3));
         }),
     );
-    eq<typeof badValSwitchMappedToNonResult, Observable<Result<Date, number>>>(true);
+    eq<typeof badValSwitchMappedToNonResult, Observable<Result<Date, number>>>(
+        true,
+    );
     expect(badValSwitchMappedToNonResult).toMatchObsResult(Err(0));
 });
 
-test('switchMap', async () => {
+test("switchMap", async () => {
     const goodValMergeMappedToResult = goodVal.pipe(
         resultMergeMap((a) => {
             eq<typeof a, string>(true);
             return of(Ok(new Date(3)) as Result<Date, string>);
         }),
     );
-    eq<typeof goodValMergeMappedToResult, Observable<Result<Date, number | string>>>(true);
+    eq<
+        typeof goodValMergeMappedToResult,
+        Observable<Result<Date, number | string>>
+    >(true);
     expect(goodValMergeMappedToResult).toMatchObsResult(Ok(new Date(3)));
 
     const goodValMergeMappedToBadResult = goodVal.pipe(
@@ -177,7 +193,10 @@ test('switchMap', async () => {
             return of(Err(new Date(3)) as Result<number, Date>);
         }),
     );
-    eq<typeof goodValMergeMappedToBadResult, Observable<Result<number, Date | number>>>(true);
+    eq<
+        typeof goodValMergeMappedToBadResult,
+        Observable<Result<number, Date | number>>
+    >(true);
     expect(goodValMergeMappedToBadResult).toMatchObsResult(Err(new Date(3)));
 
     const badValMergeMappedToResult = badVal.pipe(
@@ -186,7 +205,10 @@ test('switchMap', async () => {
             return of(Ok(new Date(3)) as Result<Date, string>);
         }),
     );
-    eq<typeof badValMergeMappedToResult, Observable<Result<Date, number | string>>>(true);
+    eq<
+        typeof badValMergeMappedToResult,
+        Observable<Result<Date, number | string>>
+    >(true);
     expect(badValMergeMappedToResult).toMatchObsResult(Err(0));
 
     const goodValMergeMappedToNonResult = goodVal.pipe(
@@ -195,7 +217,9 @@ test('switchMap', async () => {
             return of(new Date(3));
         }),
     );
-    eq<typeof goodValMergeMappedToNonResult, Observable<Result<Date, number>>>(true);
+    eq<typeof goodValMergeMappedToNonResult, Observable<Result<Date, number>>>(
+        true,
+    );
     expect(goodValMergeMappedToNonResult).toMatchObsResult(Ok(new Date(3)));
 
     const badValMergeMappedToNonResult = badVal.pipe(
@@ -204,14 +228,16 @@ test('switchMap', async () => {
             return of(new Date(3));
         }),
     );
-    eq<typeof badValMergeMappedToNonResult, Observable<Result<Date, number>>>(true);
+    eq<typeof badValMergeMappedToNonResult, Observable<Result<Date, number>>>(
+        true,
+    );
     expect(badValMergeMappedToNonResult).toMatchObsResult(Err(0));
 });
 
-test('filterResultOk', () => {
+test("filterResultOk", () => {
     const filteredGoodVal = goodVal.pipe(filterResultOk());
     eq<typeof filteredGoodVal, Observable<string>>(true);
-    expect(filteredGoodVal).toMatchObs('good');
+    expect(filteredGoodVal).toMatchObs("good");
 
     let hitBadVal = false;
     const filteredBadVal = badVal.pipe(filterResultOk());
@@ -221,7 +247,7 @@ test('filterResultOk', () => {
     expect(hitBadVal).toEqual(false);
 });
 
-test('filterResultErr', () => {
+test("filterResultErr", () => {
     let hitGoodVal = false;
     const filteredGoodVal = goodVal.pipe(filterResultErr());
     eq<typeof filteredGoodVal, Observable<number>>(true);
@@ -234,19 +260,19 @@ test('filterResultErr', () => {
     expect(filteredBadVal).toMatchObs(0);
 });
 
-test('tapResultOk', () => {
+test("tapResultOk", () => {
     let hitGoodVal = false;
     const tappedGoodVal = goodVal.pipe(
         tapResultOk((val) => {
             eq<typeof val, string>(true);
-            expect(val).toEqual('good');
+            expect(val).toEqual("good");
 
             hitGoodVal = true;
         }),
     );
 
     eq<typeof tappedGoodVal, Observable<Result<string, number>>>(true);
-    expect(tappedGoodVal).toMatchObsResult(Ok('good'));
+    expect(tappedGoodVal).toMatchObsResult(Ok("good"));
     expect(hitGoodVal).toEqual(true);
 
     let hitBadVal = false;
@@ -263,7 +289,7 @@ test('tapResultOk', () => {
     expect(hitBadVal).toEqual(false);
 });
 
-test('tapResultErr', () => {
+test("tapResultErr", () => {
     let hitGoodVal = false;
     const tappedGoodVal = goodVal.pipe(
         tapResultErr((val) => {
@@ -274,7 +300,7 @@ test('tapResultErr', () => {
     );
 
     eq<typeof tappedGoodVal, Observable<Result<string, number>>>(true);
-    expect(tappedGoodVal).toMatchObsResult(Ok('good'));
+    expect(tappedGoodVal).toMatchObsResult(Ok("good"));
     expect(hitGoodVal).toEqual(false);
 
     let hitBadVal = false;